removing links to the old pfam and updating the header
